+++ controllers/contacts/createContact.js
@@ -1,4 +1,4 @@
-const Contact = require('../../models/contactModel');
+const Contact = require('../../models');
 const { catchAsync } = require('../../utils');
 
 /**
@@ -10,7 +10,7 @@ const { catchAsync } = require('../../utils');
  * @returns {Promise<void>} - Проміс, що вирішується після створення контакту.
  * @throws {Error} - Якщо сталася помилка при створенні контакту.
  */
-exports.createContact = catchAsync(async (req, res) => {
+const createContact = catchAsync(async (req, res) => {
   const newContact = await Contact.create(req.body);
 
   newContact.password = undefined;
@@ -20,3 +20,5 @@ exports.createContact = catchAsync(async (req, res) => {
     contact: newContact,
   });
 });
+
+module.exports = createContact;
