+++ controllers/contacts/index.js
@@ -2,12 +2,14 @@ const createContact = require('./createContact');
 const getAllContacts = require('./getAllContacts');
 const getOneContact = require('./getOneContact');
 const updateContact = require('./updateContact');
+const removeContact = require('./removeContact');
 const updateStatus = require('./updateStatus');
 
 module.exports = {
   createContact,
   getAllContacts,
   getOneContact,
+  removeContact,
   updateContact,
   updateStatus,
 };
