+++ middlewares/contacts/checkBoolean.js
@@ -1,0 +1,12 @@
+const { catchAsync } = require('../../utils');
+
+exports.checkBoolean = catchAsync(async (req, res, next) => {
+  const { favorite } = req.body;
+
+  // Check if the value is boolean
+  if (typeof favorite !== 'boolean') {
+    return res.status(400).json({ error: 'The value must be boolean' });
+  }
+
+  next();
+});
