+++ middlewares/contacts/index.js
@@ -1,0 +1,9 @@
+const { checkBoolean } = require('./checkBoolean');
+const { checkContactId } = require('./checkContactId');
+const { checkCreateContactData } = require('./checkCreateContactData');
+
+module.exports = {
+  checkBoolean,
+  checkContactId,
+  checkCreateContactData,
+};
